+++ packages/seacows-amm/contracts/base/SeacowsERC721TradePairFactory.sol
@@ -12,7 +12,7 @@ contract SeacowsERC721TradePairFactory is ISeacowsERC721TradePairFactory {
     address private immutable _template;
 
     // token => collection => fee => pair
-    mapping(address => mapping(address => mapping(uint112 => address))) private _pairs;
+    mapping(address => mapping(address => mapping(uint256 => address))) private _pairs;
 
     constructor(address template_) {
         _template = template_;
@@ -22,11 +22,11 @@ contract SeacowsERC721TradePairFactory is ISeacowsERC721TradePairFactory {
         return _template;
     }
 
-    function getPair(address _token, address _collection, uint112 _fee) public view returns (address) {
+    function getPair(address _token, address _collection, uint256 _fee) public view returns (address) {
         return _pairs[_token][_collection][_fee];
     }
 
-    function _createPair(address _token, address _collection, uint112 _fee) internal returns (address _pair) {
+    function _createPair(address _token, address _collection, uint256 _fee) internal returns (address _pair) {
         require(_pairs[_token][_collection][_fee] == address(0), "Factory: Pair already exists");
         _pair = Clones.clone(template());
         ISeacowsERC721TradePair(_pair).initialize(_token, _collection, _fee);
