+++ packages/seacows-amm/contracts/interfaces/ISeacowsERC721TradePairFactory.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.13;
 
 interface ISeacowsERC721TradePairFactory {
     function template() external view returns (address);
-    function getPair(address _collection, address _token, uint112 _fee) external view returns (address);
+    function getPair(address _collection, address _token, uint256 _fee) external view returns (address);
 }
